exercice1/requirements.txt
django
djangorestframework
pandas
requests
tqdm